add action buttons menu component
